MODINVOICE-93. Update error code names
