+++ descriptors/ModuleDescriptor-template.json
@@ -49,7 +49,10 @@
           "permissionsRequired": ["invoice.invoices.item.put"],
           "permissionsDesired": [
             "invoices.acquisitions-units-assignments.manage",
-            "invoices.fiscal-year.update"
+            "invoices.fiscal-year.update",
+            "invoice.item.approve",
+            "invoice.item.pay",
+            "invoice.item.cancel"
           ],
           "modulePermissions": [
             "configuration.entries.collection.get",
