+++ src/main/java/org/folio/invoices/utils/AcqDesiredPermissions.java
@@ -10,6 +10,7 @@ public enum AcqDesiredPermissions {
   MANAGE("invoices.acquisitions-units-assignments.manage"),
   APPROVE("invoice.item.approve"),
   PAY("invoice.item.pay"),
+  CANCEL("invoice.item.cancel"),
   FISCAL_YEAR_UPDATE("invoices.fiscal-year.update");
 
   private String permission;
